# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Restful API for web app to buy air tickets
  version: 1.1.0
servers:
  - url: https://api.airtickets.com
paths:
  /flights:
    get:
      summary: Get flights with filters
      tags:
        - Flights
      operationId: findFlights
      parameters: 
        - name: departurePoint
          in: query
          required: true
          schema:
            type: string
        - name: arrivalPoint
          in: query
          required: true
          schema:
            type: string
        - name: departureDateTime
          in: query
          required: true
          schema:
            type: string
            format: 2023-08-05T15:00:00:02
      responses:
        '200':
          description: Successful response with a list of flights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flights"
        '400':
          description: Incorrect input data
    post:
      summary: Create flight
      tags:
        - Flights
      operationId: createFlight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Flight"
      responses:
        '200':
          description: Successful response with created flight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flight"
        '400':
          description: Incorrect input data
        '403':
          description: User doesn't have access
  /flights/{flightId}:
    get:
      summary: Get flight by id
      tags:
        - Flights
      operationId: getFlightById
      parameters: 
        - name: flightId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response with a flight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flight"
        '401':
          description: User is unauthorized
        '404':
          description: Flight is not in DB
    patch:
      summary: Modify flight by id
      tags:
        - Flights
      operationId: changeFlightById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Flight"
      parameters: 
        - name: flightId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response with patched flight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flight"
        '400':
          description: Incorrect input data
        '401':
          description: User is unauthorized
        '404':
          description: Flight is not in DB
    delete:
      summary: Delete flight by id
      tags:
        - Flights
      operationId: deleteFlightById
      parameters: 
        - name: flightId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Flight is deleted
        '401':
          description: User is unauthorized
        '404':
          description: Flight is not in DB          
...
  /planes:
    get:
      summary: Get all planes
      tags:
        - Planes
      operationId: getAllPlanes
      responses:
        '200':
          description: Successful response with a list of planes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Planes"
    post:
      summary: Create plane
      tags:
        - Planes
      operationId: createPlane
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Plane"
      responses:
        '200':
          description: Successful response with created plane
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plane"
        '204':
          description: Plane exists
        '400':
          description: Incorrect input data
        '401':
          description: User is unauthorized
  /planes/{planeId}:
    get:
      summary: Get plane by id
      tags:
        - Planes
      operationId: getPlaneById
      parameters:
        - name: planeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response with plane
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plane"
        '401':
          description: User is unauthorized
    patch:
      summary: Modify plane by id
      tags:
        - Planes
      operationId: changePlaneById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Plane"
      parameters: 
        - name: planeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response with created ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plane"
        '400':
          description: Incorrect input data
        '401':
          description: User is unauthorized
        '404':
          description: Plane is not in DB
    delete:
      summary: Delete plane by id
      tags:
        - Planes
      operationId: deletePlaneById
      parameters: 
        - name: planeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
        '401':
          description: User is unauthorized
        '404':
          description: Plane is not in DB
  /flights/{flightId}/planes:
    get:
      summary: Get plane by flightId
      tags:
        - Flights
      operationId: getPlaneByFlightId
      parameters:
        - name: flightId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response with plane
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plane"
        '401':
          description: User is unauthorized
...
  /orders/{orderId}:
    patch:
      summary: Modify order
      tags:
        - Orders
      operationId: changeOrder
      parameters: 
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response with modified order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        '401':
          description: User is unauthorized
components:
  schemas:
    Flight:
      type: object
      required:
        - flightId
        - planeId
        - departurePoint
        - arrivalPoint
        - departureDateTime
        - arrivalDateTime
      properties:
        flightId:
          readOnly: true
          type: integer
          example: 34
        planeId:
          type: integer
          example: 2
        departurePoint:
          type: string
          example: Москва (Шереметьево)
        arrivalPoint:
          type: string
          example: Санкт-Петербург (Пулково)
        departureDateTime:
          type: string
          format: 2023-08-05T15:00:00:02
          example: 2023-08-05T15:00:00:02
        arrivalDateTime:
          type: string
          format: 2023-08-05T15:00:00+02
          example: 2023-08-05T15:00:00+02
    Flights:
      type: array
      items:
        $ref: "#/components/schemas/Flight"
...
    Plane:
      type: object
      required: 
        - planeId
        - manufacturer
        - model
        - economyClassNum
        - businessClassNum
        - firstClassNum
      properties:
        planeId:
          type: integer
          readOnly: true
          example: 3
        manufacturer:
          type: string
          example: Боинг
        model:
          type: string
          example: 777
        economyClassNum:
          type: integer
          example: 250
        businessClassNum:
          type: integer
          example: 25
        firstClassNum:
          type: integer
          example: 10
    Planes:
      type: array
      items:
        $ref: "#/components/schemas/Plane"
...
    Order:
      type: object
      required: 
        - orderId
        - userId
        - status
      properties:
        orderId:
          readOnly: true
          type: integer
          example: 1
        userId:
          readOnly: true
          type: integer
          example: 542
        status:
          type: string
          enum:
            - создан
            - оплачен
          example: создан
...